requirements.txt
azure-common==1.1.28
azure-core==1.29.5
azure-identity==1.15.0
blinker==1.6.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.7
distlib==0.3.8
filelock==3.13.1
Flask==3.0.0
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msal==1.26.0
msal-extensions==1.1.0
msrest==0.7.1
oauthlib==3.2.2
packaging==23.2
platformdirs==4.1.0
portalocker==2.8.2
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python-http-client==3.3.7
requests==2.31.0
requests-oauthlib==1.3.1
sendgrid==6.11.0
six==1.16.0
starkbank-ecdsa==2.2.0
typing_extensions==4.8.0
urllib3==2.0.6
virtualenv==20.25.0
Werkzeug==3.0.0
msgraph-core==0.2.2